[refactor] Rename factoryadapter to v1adapter
