+++ storage_v2/v1adapter/factory_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2024 The Jaeger Authors.
 // SPDX-License-Identifier: Apache-2.0
 
-package factoryadapter
+package v1adapter
 
 import (
 	"context"
